fully working authorization and validationgit add *git add *git add *git add *!

+++ bankend_server/controllers/controller_feed.js
@@ -3,6 +3,7 @@ const Post = require('../models/post');
 const fs = require('fs');
 const path = require('path');
 const perPage = 2;
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -49,22 +50,36 @@ exports.createPost = (req, res, next) => {
 
     const imageUrl = req.file.path;
     const title = req.body.title;
-    const content = req.body.content
+    const content = req.body.content;
+    let creator;
     
     const post = new Post({
         title: title,
         content: content,
-        creator: {
-            name: 'pooper'
-        },
-        imageUrl: imageUrl
+        creator: req.userId,
+        imageUrl: imageUrl,
+
     })
 
     post.save().then(result => {
-        // console.log('post creation route:', result);
+        
+        return User.findById(req.userId);
+    })
+    .then(user => {
+
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
+    .then(result => {
+        console.log('post data', post);
         res.status(201).json({
             message: 'post created successfully',
-            post: post
+            post: post,
+            creator: {
+                _id: creator._id, 
+                name: creator.name
+            }
         })
     }).catch(err => {
         if (!err.statusCode) {
@@ -126,6 +141,13 @@ exports.editPost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+
+        if (req.userId !== post.creator.toString()) {
+            const error = new Error('Not authroized');
+            error.statusCode = 403;
+            throw error;
+        }
+
         if (post.imageUrl !== editedImageUrl) {
             clearImage(post.imageUrl)
         }
@@ -160,8 +182,20 @@ exports.deletePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if (req.userId !== post.creator.toString()) {
+            const error = new Error('Not authroized');
+            error.statusCode = 403;
+            throw error;
+        }
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);
+    })
+    .then(result => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        user.posts.pull(postId);
+        return user.save();
     })
     .then(result => {
         return res.status(200).json({
